#Zakładka 1 - ID = NULL error, change BIGSERIAL

+++ src/main/resources/db/migration/V1.002__create_author_table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE authors
 (
-    id  INT          NOT NULL PRIMARY KEY,
+    id  BIGSERIAL          NOT NULL PRIMARY KEY,
     first_name VARCHAR(255) NOT NULL,
     last_name  VARCHAR(255) NOT NULL,
     gender     VARCHAR(255) NOT NULL,
